refactor: Rename pkg cleanup to metaclenaup
