+++ internal/pkg/service/stream/storage/config.go
@@ -1,17 +1,17 @@
 package storage
 
 import (
-	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/cleanup"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/level"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/metacleanup"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/statistics"
 )
 
 // Config contains global configuration for the storage.
 type Config struct {
-	VolumesPath string            `configKey:"volumesPath" configUsage:"Mounted volumes path, each volume is in \"{type}/{label}\" subdir." validate:"required"`
-	Statistics  statistics.Config `configKey:"statistics"`
-	Cleanup     cleanup.Config    `configKey:"cleanup"`
-	Level       level.Config      `configKey:"level"`
+	VolumesPath string             `configKey:"volumesPath" configUsage:"Mounted volumes path, each volume is in \"{type}/{label}\" subdir." validate:"required"`
+	Statistics  statistics.Config  `configKey:"statistics"`
+	Cleanup     metacleanup.Config `configKey:"cleanup"`
+	Level       level.Config       `configKey:"level"`
 }
 
 type ConfigPatch struct {
@@ -21,7 +21,7 @@ type ConfigPatch struct {
 func NewConfig() Config {
 	return Config{
 		Statistics: statistics.NewConfig(),
-		Cleanup:    cleanup.NewConfig(),
+		Cleanup:    metacleanup.NewConfig(),
 		Level:      level.NewConfig(),
 	}
 }
