+++ internal/pkg/service/stream/storage/metacleanup/config.go
@@ -1,4 +1,4 @@
-package cleanup
+package metacleanup
 
 import "time"
