+++ internal/pkg/service/stream/storage/metacleanup/metacleanup.go
@@ -1,4 +1,6 @@
-package cleanup
+// Package metacleanup provides cleanup of expired file/slice metadata from DB.
+// The metadata cleanup then triggers cleanup of the physical disk files in the storage writer nodes.
+package metacleanup
 
 import (
 	"context"
@@ -48,13 +50,13 @@ func Start(d dependencies, cfg Config) error {
 	n := &Node{
 		config:    cfg,
 		clock:     d.Clock(),
-		logger:    d.Logger().WithComponent("storage.cleanup"),
+		logger:    d.Logger().WithComponent("storage.metadata.cleanup"),
 		telemetry: d.Telemetry(),
 		locks:     d.DistributedLockProvider(),
 		storage:   d.StorageRepository(),
 	}
 
-	if dist, err := d.DistributionNode().Group("storage.cleanup"); err == nil {
+	if dist, err := d.DistributionNode().Group("storage.metadata.cleanup"); err == nil {
 		n.dist = dist
 	} else {
 		return err
@@ -62,7 +64,7 @@ func Start(d dependencies, cfg Config) error {
 
 	ctx := context.Background()
 	if !n.config.Enabled {
-		n.logger.Info(ctx, "storage cleanup is disabled")
+		n.logger.Info(ctx, "local storage metadata cleanup is disabled")
 		return nil
 	}
 
@@ -85,8 +87,8 @@ func Start(d dependencies, cfg Config) error {
 		defer ticker.Stop()
 
 		for {
-			if err := n.cleanFiles(ctx); err != nil && !errors.Is(err, context.Canceled) {
-				n.logger.Errorf(ctx, `storage cleanup failed: %s`, err)
+			if err := n.cleanMetadata(ctx); err != nil && !errors.Is(err, context.Canceled) {
+				n.logger.Errorf(ctx, `local storage metadata cleanup failed: %s`, err)
 			}
 
 			select {
@@ -101,9 +103,9 @@ func Start(d dependencies, cfg Config) error {
 	return nil
 }
 
-// cleanFiles iterates all files and deletes the expired ones.
-func (n *Node) cleanFiles(ctx context.Context) (err error) {
-	ctx, span := n.telemetry.Tracer().Start(ctx, "keboola.go.stream.model.cleanup.node.cleanFiles")
+// cleanMetadata iterates all files and deletes the expired ones.
+func (n *Node) cleanMetadata(ctx context.Context) (err error) {
+	ctx, span := n.telemetry.Tracer().Start(ctx, "keboola.go.stream.model.cleanup.metadata.cleanMetadata")
 	defer span.End(&err)
 
 	// Measure count of deleted files
@@ -157,7 +159,7 @@ func (n *Node) cleanFile(ctx context.Context, file model.File) (err error, delet
 	ctx = ctxattr.ContextWith(ctx, attrs...)
 
 	// Trace each file
-	ctx, span := n.telemetry.Tracer().Start(ctx, "keboola.go.stream.model.cleanup.node.cleanFile")
+	ctx, span := n.telemetry.Tracer().Start(ctx, "keboola.go.stream.model.cleanup.metadata.cleanFile")
 	defer span.End(&err)
 
 	// Check if the file is expired
