+++ internal/pkg/service/stream/storage/node/coordinator/coordinator.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/config"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/dependencies"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/cleanup"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/metacleanup"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/node/coordinator/fileimport"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/node/coordinator/filerotation"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/node/coordinator/slicerotation"
@@ -30,7 +30,7 @@ func Start(ctx context.Context, d dependencies.CoordinatorScope, cfg config.Conf
 		return err
 	}
 
-	if err := cleanup.Start(d, cfg.Storage.Cleanup); err != nil {
+	if err := metacleanup.Start(d, cfg.Storage.Cleanup); err != nil {
 		return err
 	}
